+++ .vscode/PythonImportHelper-v2-Completion.json
@@ -1307,12 +1307,28 @@
         "documentation": {}
     },
     {
+        "label": "Ui_Dialog",
+        "importPath": "ui_files.ui_password_entry",
+        "description": "ui_files.ui_password_entry",
+        "isExtraImport": true,
+        "detail": "ui_files.ui_password_entry",
+        "documentation": {}
+    },
+    {
         "label": "logging",
         "kind": 6,
         "isExtraImport": true,
         "importPath": "logging",
         "description": "logging",
         "detail": "logging",
+        "documentation": {}
+    },
+    {
+        "label": "RotatingFileHandler",
+        "importPath": "logging.handlers",
+        "description": "logging.handlers",
+        "isExtraImport": true,
+        "detail": "logging.handlers",
         "documentation": {}
     },
     {
@@ -1365,19 +1381,11 @@
         "documentation": {}
     },
     {
-        "label": "MessageManager",
-        "importPath": "utils.message_manager",
-        "description": "utils.message_manager",
+        "label": "datetime",
+        "importPath": "datetime",
+        "description": "datetime",
         "isExtraImport": true,
-        "detail": "utils.message_manager",
-        "documentation": {}
-    },
-    {
-        "label": "MessageManager",
-        "importPath": "utils.message_manager",
-        "description": "utils.message_manager",
-        "isExtraImport": true,
-        "detail": "utils.message_manager",
+        "detail": "datetime",
         "documentation": {}
     },
     {
@@ -1551,6 +1559,14 @@
         "documentation": {}
     },
     {
+        "label": "MessageManager",
+        "importPath": "utils.message_manager",
+        "description": "utils.message_manager",
+        "isExtraImport": true,
+        "detail": "utils.message_manager",
+        "documentation": {}
+    },
+    {
         "label": "traceback",
         "kind": 6,
         "isExtraImport": true,
@@ -1679,28 +1695,28 @@
     {
         "label": "get_log_path",
         "kind": 2,
-        "importPath": "utils.global_vars",
-        "description": "utils.global_vars",
+        "importPath": "utils.logging_config",
+        "description": "utils.logging_config",
         "peekOfCode": "def get_log_path() -> str:\n    \"\"\"Get the appropriate log file path whether running as script or frozen exe\"\"\"\n    if getattr(sys, 'frozen', False):\n        # Running as compiled executable\n        base_path = os.path.dirname(sys.executable)\n    else:\n        # Running as script\n        base_path = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))\n    # Create logs directory if it doesn't exist\n    logs_dir = os.path.join(base_path, 'logs')",
-        "detail": "utils.global_vars",
+        "detail": "utils.logging_config",
         "documentation": {}
     },
     {
         "label": "setup_logger",
         "kind": 2,
-        "importPath": "utils.global_vars",
-        "description": "utils.global_vars",
-        "peekOfCode": "def setup_logger() -> logging.Logger:\n    \"\"\"Setup and configure the logger\n    Returns:\n        logging.Logger: Configured logger instance\n    \"\"\"\n    # Create logger\n    logger = logging.getLogger('multipack_parser')\n    logger.setLevel(logging.INFO)  # Set level first\n    # Remove any existing handlers\n    logger.handlers.clear()",
-        "detail": "utils.global_vars",
+        "importPath": "utils.logging_config",
+        "description": "utils.logging_config",
+        "peekOfCode": "def setup_logger() -> logging.Logger:\n    \"\"\"Setup and configure the logger\"\"\"\n    logger = logging.getLogger('multipack_parser')\n    logger.setLevel(logging.INFO)\n    # Remove any existing handlers\n    logger.handlers.clear()\n    try:\n        log_formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')\n        # File handler\n        log_path = get_log_path()",
+        "detail": "utils.logging_config",
         "documentation": {}
     },
     {
         "label": "logger",
         "kind": 5,
-        "importPath": "utils.global_vars",
-        "description": "utils.global_vars",
-        "peekOfCode": "logger = setup_logger()\n# Test the logger immediately\nlogger.info(\"Global variables module initialized\")\nprocess: Optional[multiprocessing.Process] = None\n#######################################\n# Settings\nsettings: Optional['Settings'] = None\n# Variables\nsettings_file: str = 'MultipackParser.conf'\nVERSION: str = '1.5.7'",
-        "detail": "utils.global_vars",
+        "importPath": "utils.logging_config",
+        "description": "utils.logging_config",
+        "peekOfCode": "logger = setup_logger()",
+        "detail": "utils.logging_config",
         "documentation": {}
     },
     {
@@ -1731,29 +1747,11 @@
         "documentation": {}
     },
     {
-        "label": "logger",
-        "kind": 5,
-        "importPath": "utils.message_manager",
-        "description": "utils.message_manager",
-        "peekOfCode": "logger = global_vars.logger\nclass MessageManager:\n    \"\"\"Manager for messages.\n    \"\"\"\n    def __init__(self):\n        \"\"\"Initialize the message manager.\n        \"\"\"\n        logger.debug(\"Initializing MessageManager\")\n        self._messages: List[Message] = []\n        self._active_messages: List[Message] = []",
-        "detail": "utils.message_manager",
-        "documentation": {}
-    },
-    {
         "label": "Settings",
         "kind": 6,
         "importPath": "utils.settings",
         "description": "utils.settings",
-        "peekOfCode": "class Settings:\n    \"\"\"Settings class.\n    \"\"\"\n    def __init__(self):\n        \"\"\"Initialize the settings with QSettings.\n        \"\"\"\n        # Create QSettings instance with organization and application name\n        self.qsettings = QSettings(\"Multipack\", \"MultipackParser\")\n        # Define default settings\n        self.default_settings = {",
-        "detail": "utils.settings",
-        "documentation": {}
-    },
-    {
-        "label": "logger",
-        "kind": 5,
-        "importPath": "utils.settings",
-        "description": "utils.settings",
-        "peekOfCode": "logger = global_vars.logger\nclass Settings:\n    \"\"\"Settings class.\n    \"\"\"\n    def __init__(self):\n        \"\"\"Initialize the settings with QSettings.\n        \"\"\"\n        # Create QSettings instance with organization and application name\n        self.qsettings = QSettings(\"Multipack\", \"MultipackParser\")\n        # Define default settings",
+        "peekOfCode": "class Settings:\n    \"\"\"Settings class.\n    \"\"\"\n    def __init__(self):\n        \"\"\"Initialize the settings with QSettings.\n        \"\"\"\n        logger.debug(\"Initializing Settings\")\n        # Create QSettings instance with organization and application name\n        self.qsettings = QSettings(\"Multipack\", \"MultipackParser\")\n        # Define default settings",
         "detail": "utils.settings",
         "documentation": {}
     },
